requirements.txt
openai>=1.0.0
pyaudio>=0.2.11
numpy>=1.23
soundfile>=0.12
python-dotenv>=1.0.0
